backend/requirements.txt
fastapi
uvicorn
python-dotenv
carconnectivity
